feat: started process of replacing monstrous trekpanel with something similar
